Refactor: sort hooks into subfolders
