+++ src/hooks/index.ts
@@ -1,9 +1,11 @@
-export { useAddCardToDeck } from "./useAddCardToDeck";
-export { useAddDeck } from "./useAddDeck";
-export { useAuthenticateUser } from "./useAuthenticateUser";
-export { useAxios } from "./useAxios";
-export { useCreateCard } from "./useCreateCard";
-export { useGameCards } from "./useGameCards";
-export { useGetData } from "./useGetData";
-export { usePlayers } from "./usePlayers";
-export { useSetupDeck } from "./useSetupDeck";
+export { useAxios } from "./api/useAxios";
+export { useGetData } from "./api/useGetData";
+
+export { useAddCardToDeck } from "./forms/useAddCardToDeck";
+export { useAddDeck } from "./forms/useAddDeck";
+export { useAuthenticateUser } from "./forms/useAuthenticateUser";
+export { useCreateCard } from "./forms/useCreateCard";
+
+export { useGameCards } from "./game/useGameCards";
+export { usePlayers } from "./game/usePlayers";
+export { useSetupDeck } from "./game/useSetupDeck";
